agent/pyproject.toml
[project]
name = "career-bloom-agents"
version = "0.1.0"
description = "CareerBloomAgent ADK multi-agent system"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.3.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-cloud-firestore>=2.16.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
